Cache removal + shared call hierarchy engine extraction
