+++ analysis-api/src/main/kotlin/io/github/amichne/kast/api/CallHierarchyQuery.kt
@@ -10,5 +10,4 @@ data class CallHierarchyQuery(
     val maxTotalCalls: Int = 256,
     val maxChildrenPerNode: Int = 64,
     val timeoutMillis: Long? = null,
-    val persistToGitShaCache: Boolean = false,
 )
