+++ analysis-api/src/main/kotlin/io/github/amichne/kast/api/CallHierarchyResult.kt
@@ -6,7 +6,6 @@ import kotlinx.serialization.Serializable
 data class CallHierarchyResult(
     val root: CallNode,
     val stats: CallHierarchyStats,
-    val persistence: CallHierarchyPersistence? = null,
     val schemaVersion: Int = SCHEMA_VERSION,
 )
 
@@ -20,11 +19,4 @@ data class CallHierarchyStats(
     val maxTotalCallsReached: Boolean,
     val maxChildrenPerNodeReached: Boolean,
     val filesVisited: Int,
-)
-
-@Serializable
-data class CallHierarchyPersistence(
-    val gitSha: String,
-    val cacheFilePath: String,
-    val cacheHit: Boolean,
 )
